Fixed skip layer bug

+++ dd/Definitions.hpp
@@ -23,7 +23,7 @@ using RefCount = std::uint32_t;
 static_assert(std::is_unsigned_v<RefCount>, "RefCount should be unsigned.");
 
 // floating point type to use
-using fp = float;//double;
+using fp = double;//double;
 static_assert(
     std::is_floating_point_v<fp>,
     "fp should be a floating point type (float, *double*, long double)");
